move test user on app.jsx in home.jsx
